+++ src/main/java/com/coolerpromc/uncrafteverything/block/custom/AutoUncraftingTableBlock.java
@@ -8,7 +8,9 @@ import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.*;
@@ -58,6 +60,19 @@ public class AutoUncraftingTableBlock extends BaseEntityBlock {
             }
         }
         super.neighborChanged(state, level, pos, neighborBlock, neighborPos, movedByPiston);
+    }
+
+    @Override
+    public void setPlacedBy(Level level, BlockPos pos, BlockState state, @Nullable LivingEntity pPlacer, ItemStack pStack) {
+        if (!level.isClientSide()){
+            boolean powered = level.hasNeighborSignal(pos);
+            BlockEntity entity = level.getBlockEntity(pos);
+            if (entity instanceof AutoUncraftingTableBlockEntity blockEntity){
+                blockEntity.setActive(powered);
+                level.setBlock(pos, state.setValue(POWERED, powered), 3);
+            }
+        }
+        super.setPlacedBy(level, pos, state, pPlacer, pStack);
     }
 
     @Override